+++ packages/icons/src/lib/icons/CheckCircle.tsx
@@ -6,17 +6,27 @@ import { SVGRProps } from '../type';
 const SvgCheckCircle = ({
     title,
     titleId,
+    isResponsive = false,
+    style,
     ...props
 }: SVGProps<SVGSVGElement> & SVGRProps) => (
     <svg
         width="1em"
         height="1em"
         viewBox="0 0 24 24"
-        fill="none"
+        fill="currentColor"
         xmlns="http://www.w3.org/2000/svg"
         role="img"
         data-testid="test-icon"
         aria-hidden="true"
+        style={
+            isResponsive
+                ? style
+                : {
+                      flex: 'none',
+                      ...style,
+                  }
+        }
         aria-labelledby={titleId}
         {...props}
     >
@@ -25,13 +35,11 @@ const SvgCheckCircle = ({
             fillRule="evenodd"
             clipRule="evenodd"
             d="M19.364 12a7.364 7.364 0 1 0-14.728 0 7.364 7.364 0 0 0 14.728 0ZM12 3a9 9 0 1 1 0 18 9 9 0 0 1 0-18Z"
-            fill="#000"
         />
         <path
             fillRule="evenodd"
             clipRule="evenodd"
             d="M15.873 9a.886.886 0 0 1 0 1.255l-4.5 4.491a.89.89 0 0 1-1.258 0l-1.988-1.984a.886.886 0 0 1 0-1.256.89.89 0 0 1 1.258 0l1.36 1.357L14.614 9a.89.89 0 0 1 1.258 0Z"
-            fill="#000"
         />
     </svg>
 );